Replace `OnDomEvents.instead()` and `.passive() with `handleDomEvents()` action
