+++ src/dom/actions/index.ts
@@ -1,1 +1,2 @@
 export * from './capture-dom-events';
+export * from './handle-dom-events';
